requirements.txt
better_bing_image_downloader==1.1.2
Pillow==10.4.0
webp==0.4.0
augmentor== 0.2.12
